implement invite command and it seems to work, also fixed wat i think ncotte forgot to push

+++ inc/User.h
@@ -6,20 +6,24 @@
 /*   By: shalimi <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/11 16:07:33 by shalimi           #+#    #+#             */
-/*   Updated: 2023/03/23 17:03:09 by shalimi          ###   ########.fr       */
+/*   Updated: 2023/03/27 17:08:50 by shalimi          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 
 #ifndef USER_H
 # define USER_H
-
 # include "ft_irc.hpp"
 # include "Operator.hpp"
-
+# include <vector>
+# include <netinet/in.h>
 //# include <stdio.h>
 # include <cstring>
+# include <iostream>
+# include <sstream>
+# include <sys/socket.h>
 # include <arpa/inet.h>
+# include <unistd.h>
 
 # define PASS_FLAG				0x00000001
 # define NICK_FLAG				0x00000002
@@ -46,55 +50,54 @@ class User
 		~User();
 
 
-		void					sendReply(std::string const & reply) const;
+		void				sendReply(std::string const & reply) const;
 
-		User &					operator=(User const & src);
-		bool					operator==(User const & src);
+		User &	operator=(User const & src);
+		bool	operator==(User const & src);
 
-		void					welcome(Server const & server) const;
-
-		void					handleCmd(Server & server);
+		void				welcome(Server const & server) const;
+		void				handleCmd(Server & server);
 
 		/* Setters */
-		void					addChannel(Channel * chan);
-		void					removeChannel(Channel * chan);
-		void					setKEvent(struct kevent * event);
-		void					auth();
-		void					setFd(int i);
-		void					setUsername(std::string const &user);
-		void					setRealname(std::string const &user);
-		void					setNickname(std::string const &nick);
-		void					setHostname();
-		void					setAway(bool flag);
-		void					setInvisible(bool flag);
-		void					setWallops(bool flag);
-		void					setRestricted(bool flag);
-		void					setGlobalOperator(bool flag);
-		void					setLocalOperator(bool flag);
-		bool					loginOperator(Operator const * op, std::string const &password);
+		void				addChannel(Channel * chan);
+		void				removeChannel(Channel * chan);
+		void				setKEvent(struct kevent * event);
+		void				auth();
+		void				setFd(int i);
+		void				setUsername(std::string const &user);
+		void				setRealname(std::string const &user);
+		void				setNickname(std::string const &nick);
+		void				setHostname();
+		void				setAway(bool flag);
+		void				setInvisible(bool flag);
+		void				setWallops(bool flag);
+		void				setRestricted(bool flag);
+		void				setGlobalOperator(bool flag);
+		void				setLocalOperator(bool flag);
+		bool				loginOperator(Operator const * op, std::string const &password);
 
 		/* Getters */
-		struct kevent *			getKEvent() const;
-		bool					hasNick() const;
-		bool					hasUser() const;
-		bool					isLog() const;
-		int &					getFd();
-		socklen_t *				getSocklenPtr();
+		struct kevent *		getKEvent() const;
+		bool				hasNick() const;
+		bool				hasUser() const;
+		bool				isLog() const;
+		int &				getFd();
+		socklen_t *			getSocklenPtr();
 		struct sockaddr_in *	getAddressPtr();
 		std::vector<Channel *>*	getChannels() const;
-		std::string 			getHostname() const;
-		std::string 			getRealname() const;
-		std::string 			getNickname() const;
-		std::string				getUsername() const;
-		std::string				prefix() const;
-		bool					hasPass() const;
-		bool					isAway() const;
-		bool					isInvisible() const;
-		bool					isWallops() const;
-		bool					isRestricted() const;
-		bool					isGlobalOperator() const;
-		bool					isLocalOperator() const;
-		bool					isOperator() const;
+		std::string getHostname() const;
+		std::string getRealname() const;
+		std::string getNickname() const;
+		std::string	getUsername() const;
+		std::string	prefix() const;
+		bool				hasPass() const;
+		bool				isAway() const;
+		bool				isInvisible() const;
+		bool				isWallops() const;
+		bool				isRestricted() const;
+		bool				isGlobalOperator() const;
+		bool				isLocalOperator() const;
+		bool				isOperator() const;
 
 	private:
 		unsigned int 			_boolFlags;
@@ -108,7 +111,6 @@ class User
 		struct kevent *			_event;
 		std::vector<Channel *>*	_channels;
 		char					_buffer[513];
-		size_t					_bufferLength;
+		int						_bufferLength;
 };
-
 #endif
